ting bare funker

+++ Store/StarDataStore.cs
@@ -7,8 +7,18 @@ public static class StarDataStore
 {
 
 
-    public static async Task<StarItem?> AddStarItem(ItemCacheDb db, StarItem item)
+    public static async Task<StarItem?> AddStarItem(ItemCacheDb db, StarItem item, string username)
     {
+        var user = await db.Users.FirstOrDefaultAsync(u => u.Username == username);
+        if (user == null)
+        {
+            Console.WriteLine($"Invalid User when adding item");
+            return null;
+        } else
+        {
+            item.Username = username;
+        }
+
 
         var location = await db.StarLocations.FirstOrDefaultAsync(loc => loc.Id == item.LocationId);
         if (location == null)
@@ -33,13 +43,20 @@ public static class StarDataStore
 
     }
 
-    public static async Task<StarItem> UpdateStarItem(ItemCacheDb db, int id, StarItem item)
+    public static async Task<StarItem> UpdateStarItem(ItemCacheDb db, int id, StarItem item, string username)
     {
         StarItem? existingItem = await db.PersonalItems.FindAsync(id);
         if (existingItem == null)
         {
             throw new KeyNotFoundException("Item not found");
         }
+
+        if (existingItem.Username != username)
+        {
+            throw new UnauthorizedAccessException("You do not have permission to update this item");
+        }
+
+
 
         item.Id = id;
 
@@ -52,11 +69,17 @@ public static class StarDataStore
         return existingItem;
     }
 
-    public static async Task<bool> DeleteStarItem(ItemCacheDb db, int id)
+    public static async Task<bool> DeleteStarItem(ItemCacheDb db, int id, string username)
     {
         StarItem? existingItem = await db.PersonalItems.FindAsync(id);
         if (existingItem != null)
         {
+
+            if (existingItem.Username != username)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to update this item");
+            }
+
             db.PersonalItems.Remove(existingItem);
             await db.SaveChangesAsync();
             return true;
@@ -64,9 +87,9 @@ public static class StarDataStore
         return false;
     }
 
-    public static async Task<List<StarItem>> GetPersonalItems(ItemCacheDb db)
+    public static async Task<List<StarItem>> GetPersonalItems(ItemCacheDb db, string username)
     {
-        List<StarItem> items = await db.PersonalItems.ToListAsync();
+        List<StarItem> items = await db.PersonalItems.Where(user => user.Username.Equals(username)).ToListAsync();
         return items;
     }
